+++ antareslauncher/use_cases/retrieve/clean_remote_server.py
@@ -2,7 +2,9 @@ import copy
 from pathlib import Path
 
 from antareslauncher.display.idisplay import IDisplay
-from antareslauncher.remote_environnement import iremote_environment
+from antareslauncher.remote_environnement.remote_environment_with_slurm import (
+    RemoteEnvironmentWithSlurm,
+)
 from antareslauncher.study_dto import StudyDTO
 
 
@@ -13,7 +15,7 @@ class RemoteServerNotCleanException(Exception):
 class RemoteServerCleaner:
     def __init__(
         self,
-        env: iremote_environment.IRemoteEnvironment,
+        env: RemoteEnvironmentWithSlurm,
         display: IDisplay,
     ):
         self._display = display