+++ CHANGES.md
@@ -2,6 +2,14 @@
 
 ## [1.3.0] - (unreleased)
 
+## [1.2.3] - 2023-03-16
+
+### Fixes
+
+- Correct download progress bar in logs [#40](https://github.com/AntaresSimulatorTeam/antares-launcher/pull/40)
+
+- Correct SLURM job status checking [#43](https://github.com/AntaresSimulatorTeam/antares-launcher/pull/43)
+
 ## [1.2.2] - 2023-03-02
 
 ### Fixes
@@ -76,7 +84,9 @@
 - Remove unnecessary Optional
 - Enable ssh_config_file to be `None`
 
-[1.3.0]: https://github.com/AntaresSimulatorTeam/antares-launcher/compare/v1.2.2...HEAD
+[1.3.0]: https://github.com/AntaresSimulatorTeam/antares-launcher/compare/v1.2.3...HEAD
+
+[1.2.3]: https://github.com/AntaresSimulatorTeam/antares-launcher/releases/tag/v1.2.3
 
 [1.2.2]: https://github.com/AntaresSimulatorTeam/antares-launcher/releases/tag/v1.2.2